Move program memory to seperate file
